Refactor imports to use BlktModelTypes from blanket_variables for consistency
