+++ process/models/blankets/blanket_library.py
@@ -17,6 +17,7 @@ from process.data_structure import (
     physics_variables,
     primary_pumping_variables,
 )
+from process.data_structure.blanket_variables import BlktModelTypes
 from process.models.build import FwBlktVVShape
 from process.models.engineering.ivc_functions import (
     calculate_pipe_bend_radius,
@@ -46,13 +47,6 @@ logger = logging.getLogger(__name__)
 # OB          Outboard
 # HCD         Heating & Current Drive
 # FCI         Flow Channel Insert
-
-
-class BlktModelTypes(IntEnum):
-    """Enum for blanket model types. `i_blanket_type`"""
-
-    CCFE_HCPB = 1
-    DCLL = 5
 
 
 class FWBlktCoolantLoopTypes(IntEnum):
